Initial import.
This is a refactor of the old Architecture 101 demo, which was extracted from a chapter of the courseware by the same name.
The changes are to repackage into the new subproject namespace, and to update the startup mechanism to use the new, simpler idiom of placing a startup method on the facade and simply calling it from the creationComplete of the application.